refactor: move mixin classes, also bump shadow plugin to version 8.3.3, remove redundant code from build-mod.gradle.kts
