+++ gradle.properties
@@ -3,7 +3,7 @@
 
 # Versions
 version_kotlin=2.0.21
-version_shadow=8.1.1
+version_shadow=8.3.3
 version_DevAuth=1.2.1
 version_OneConfig=0.2.2-alpha+
 version_idea_ext=1.1.10
