+++ src/main/java/net/llvg/exec/mixin/mixin/MixinItemRenderer.java
@@ -17,27 +17,22 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.llvg.exec.mixin;
+package net.llvg.exec.mixin.mixin;
 
 import net.llvg.exec.features.freecam.FreeCam;
 import net.minecraft.client.entity.AbstractClientPlayer;
-import net.minecraft.client.renderer.entity.RenderPlayer;
-import net.minecraft.client.renderer.entity.RendererLivingEntity;
+import net.minecraft.client.renderer.ItemRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.ModifyVariable;
 
-@Mixin (RenderPlayer.class)
-public abstract class MixinRenderPlayer extends RendererLivingEntity<AbstractClientPlayer> {
-        private MixinRenderPlayer() {
-                super(null, null, 0);
-        }
-        
-        @Redirect (method = "doRender(Lnet/minecraft/client/entity/AbstractClientPlayer;DDDFF)V", at = @At (value = "INVOKE", target = "Lnet/minecraft/client/entity/AbstractClientPlayer;isUser()Z"))
-        private boolean doRenderRedirect(AbstractClientPlayer instance) {
+@Mixin (ItemRenderer.class)
+public abstract class MixinItemRenderer {
+        @ModifyVariable (method = "renderItemInFirstPerson", at = @At ("STORE"), index = 3)
+        private AbstractClientPlayer renderItemInFirstPersonModifyVariable(AbstractClientPlayer abstractclientplayer) {
                 if (FreeCam.isEnabled()) {
-                        return instance == FreeCam.getCamera();
+                        return FreeCam.getCamera();
                 }
-                return instance.isUser();
+                return abstractclientplayer;
         }
 }
