+++ src/main/java/net/llvg/exec/mixin/mixin/MixinMinecraft.java
@@ -17,7 +17,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.llvg.exec.mixin;
+package net.llvg.exec.mixin.mixin;
 
 import net.llvg.exec.event.ExeCEventManager;
 import net.llvg.exec.event.events.WorldLoadEvent;
@@ -27,7 +27,7 @@ import net.minecraft.client.multiplayer.WorldClient;
 import net.minecraft.client.settings.GameSettings;
 import net.minecraft.profiler.IPlayerUsage;
 import net.minecraft.util.IThreadListener;
-import org.spongepowered.asm.lib.Opcodes;
+import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
@@ -38,9 +38,11 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin (Minecraft.class)
 public abstract class MixinMinecraft implements IThreadListener, IPlayerUsage {
+        @Shadow
+        public GameSettings gameSettings;
+        
         @Unique
         private static int exec$gameSettings$thirdPersonView$storage;
-        @Shadow public GameSettings gameSettings;
         
         @Inject (method = "clickMouse", at = @At ("HEAD"), cancellable = true)
         private void clickMouseInject(CallbackInfo ci) {
