+++ src/main/java/net/llvg/exec/mixin/mixin/MixinRenderPlayer.java
@@ -17,22 +17,27 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.llvg.exec.mixin;
+package net.llvg.exec.mixin.mixin;
 
 import net.llvg.exec.features.freecam.FreeCam;
 import net.minecraft.client.entity.AbstractClientPlayer;
-import net.minecraft.client.renderer.ItemRenderer;
+import net.minecraft.client.renderer.entity.RenderPlayer;
+import net.minecraft.client.renderer.entity.RendererLivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.Redirect;
 
-@Mixin (ItemRenderer.class)
-public abstract class MixinItemRenderer {
-        @ModifyVariable (method = "renderItemInFirstPerson", at = @At ("STORE"), index = 3)
-        private AbstractClientPlayer renderItemInFirstPersonModifyVariable(AbstractClientPlayer abstractclientplayer) {
+@Mixin (RenderPlayer.class)
+public abstract class MixinRenderPlayer extends RendererLivingEntity<AbstractClientPlayer> {
+        private MixinRenderPlayer() {
+                super(null, null, 0);
+        }
+        
+        @Redirect (method = "doRender(Lnet/minecraft/client/entity/AbstractClientPlayer;DDDFF)V", at = @At (value = "INVOKE", target = "Lnet/minecraft/client/entity/AbstractClientPlayer;isUser()Z"))
+        private boolean doRenderRedirect(AbstractClientPlayer instance) {
                 if (FreeCam.isEnabled()) {
-                        return FreeCam.getCamera();
+                        return instance == FreeCam.getCamera();
                 }
-                return abstractclientplayer;
+                return instance.isUser();
         }
 }
